adjusted the update function and tested
