+++ public/index.html
@@ -11,6 +11,7 @@
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
     <link href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" rel="stylesheet">
     <script src="js/ReservationWind.js"></script>
+    <script src="js/update.js"></script>
 </head>
 
 <body onload="viewReservation()">
